decide the bit width of all types

+++ src/build.h
@@ -12,14 +12,14 @@ typedef struct {
 	bool   debug;
 	sds    target;
 	sds    target_dir;
-	int    jobs;
+	i64    jobs;
 	sds   *features;
 	size_t features_count;
 	bool   all_features;
 	bool   no_default_features;
 } build_opts_t;
 
-int build_project(manifest_t *manifest, build_opts_t *opts);
-int build_run(manifest_t *manifest, build_opts_t *opts, sds *args, int argc);
+i64 build_project(manifest_t *manifest, build_opts_t *opts);
+i64 build_run(manifest_t *manifest, build_opts_t *opts, sds *args, i64 argc);
 
 #endif
